Added a services.c file.  This will contain the functions which correspond to each type of request from the server or client.  For example, the client's processing for sending a file, and the client's processing for receiving it.
Basically what happens is this: 
-start the server by clicking on the server button.  When you do this, it sets the behavior flag in the connectinfo structure to denote that it is a server.
-for the client, click client, which also sets the behavior flag.  Then click on a request type (download, stream, upload, multicast).  You MUST choose one of these options since there's no error handling for not choosing a setting yet.
-Then when you click connect on the client, it calls the connect function as normal, which the server accepts.  It then automatically sends an initial packet in text format which denotes what the client's request flag is set to.  You can see this in writeTCPSocket() in the socket.c file.  Right above that function is the server's handle function for when it receives this first packet (sockRead).  It will do a strcmp to the packet, and call the corresponding function to handle that type of connection in the services.c file.
-Basically what needs to be written (Jerrod) are these services.c functions, I want upload and download completely working without errors, and when you choose a streaming request, it should create a UDP socket & call a function to service the request (I will do the streaming request functions)).  For now, you can just close the socket & start over once a request has finished servicing.  The client will have to reconnect to get a new request serviced.  There's a couple of these services.c functions that are partially written, but do not work.  At one point I ended up sending a 2GB file to myself.

Here is where the problem comes in: these services functions will be receiving data for something, but this is supposed to be asyncronous, so when we get in data, it will be trying to read from both the service function and as a message through WM_READ if we have a second, idle thread.  So should we have the data processing directly in the WM_READ section and use flags in the connectInfo structure to figure out what we should be doing with the data?  I don't really care which approach is chosen, as long as it works.  Figure it out!!11  And whatever approach is taken, make sure that it doesn't interfere with the streaming audio part (ie. the flags need to work properly so that we can call a services.c function for the streaming audio section.  This probably won't effect much, but it needs to be kept in mind just in case something funky goes down).